fix: refactor component usage within context
