+++ src/components/alert-dialog/index.ts
@@ -1,3 +1,3 @@
 export { AlertDialog } from './AlertDialog'
 
-export { useAlert, AlertProvider } from './useAlert'
+export { useAlert, AlertProvider } from './alert-context/useAlert'
